tests_dev/requirements-test.txt
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist==3.3.1
httpx==0.24.1
fastapi[all]==0.100.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.19
alembic==1.11.1
factory-boy==3.3.0
freezegun==1.2.2
